refactor: Rename DriveView into DriveFolderView
